[VUFIND-1455] Move record driver traits into Feature namespace.
